Change TrainingStatus to enumeration in domain model
Change training status type to smallint

+++ src/Smart.FA.Catalog.Core/Domain/Training/Dto/TrainingDto.cs
@@ -2,7 +2,7 @@ namespace Core.Domain.Dto;
 
 public record TrainingDto
     (int TrainingId,
-    int StatusId,
+    short StatusId,
     string Title,
     string? Goal,
     string Language);
